requirements.txt
PyMuPDF
pandas
python-dotenv
PyPDF2
pytest
peewee
psycopg2-binary
python-dotenv